# backend/src/main/java/com/learn/turnup/apis/LessonsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.9.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.learn.turnup.apis;

import java.util.UUID;

import com.learn.turnup.dto.LessonDTO;
import com.learn.turnup.dto.NewLessonDTO;
import com.learn.turnup.dto.NewWordDTO;
import com.learn.turnup.dto.WordDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-08-05T10:53:27.471917+02:00[Europe/Prague]", comments = "Generator version: 7.9.0")
@Validated
@Tag(name = "Lesson", description = "Lessons own Words")
public interface LessonsApi {

    /**
     * POST /lessons/{lessonId}/copy : Copy a shared lesson
     *
     * @param xUserId  (required)
     * @param lessonId  (required)
     * @return Lesson copied (status code 200)
     *         or Invalid input (status code 400)
     *         or User ID is not recognized (status code 401)
     *         or Unauthorized access to lesson (status code 403)
     *         or Lesson not found (status code 404)
     *         or Lesson name already exists (status code 409)
     */
    @Operation(
        operationId = "copySharedLesson",
        summary = "Copy a shared lesson",
        tags = { "Lesson" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Lesson copied", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LessonDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid input"),
            @ApiResponse(responseCode = "401", description = "User ID is not recognized"),
            @ApiResponse(responseCode = "403", description = "Unauthorized access to lesson"),
            @ApiResponse(responseCode = "404", description = "Lesson not found"),
            @ApiResponse(responseCode = "409", description = "Lesson name already exists")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/lessons/{lessonId}/copy",
        produces = { "application/json" }
    )
    
    ResponseEntity<LessonDTO> copySharedLesson(
        @NotNull @Parameter(name = "X-User-Id", description = "", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "X-User-Id", required = true) UUID xUserId,
        @Parameter(name = "lessonId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("lessonId") UUID lessonId
    );


    /**
     * POST /lessons : Create a new lesson
     *
     * @param xUserId  (required)
     * @param newLessonDTO  (required)
     * @return Lesson created (status code 200)
     *         or Invalid input (status code 400)
     *         or User ID is not recognized (status code 401)
     *         or Lesson name already exists (status code 409)
     */
    @Operation(
        operationId = "createLesson",
        summary = "Create a new lesson",
        tags = { "Lesson" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Lesson created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LessonDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid input"),
            @ApiResponse(responseCode = "401", description = "User ID is not recognized"),
            @ApiResponse(responseCode = "409", description = "Lesson name already exists")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/lessons",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<LessonDTO> createLesson(
        @NotNull @Parameter(name = "X-User-Id", description = "", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "X-User-Id", required = true) UUID xUserId,
        @Parameter(name = "NewLessonDTO", description = "", required = true) @Valid @RequestBody NewLessonDTO newLessonDTO
    );


    /**
     * POST /lessons/{lessonId}/words : Create new word in a lesson
     *
     * @param xUserId  (required)
     * @param lessonId  (required)
     * @param newWordDTO  (required)
     * @return Word created (status code 200)
     *         or Invalid input (status code 400)
     *         or User ID is not recognized (status code 401)
     *         or Unauthorized access to lesson (status code 403)
     *         or Lesson not found (status code 404)
     */
    @Operation(
        operationId = "createWord",
        summary = "Create new word in a lesson",
        tags = { "Lesson" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Word created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = WordDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid input"),
            @ApiResponse(responseCode = "401", description = "User ID is not recognized"),
            @ApiResponse(responseCode = "403", description = "Unauthorized access to lesson"),
            @ApiResponse(responseCode = "404", description = "Lesson not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/lessons/{lessonId}/words",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<WordDTO> createWord(
        @NotNull @Parameter(name = "X-User-Id", description = "", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "X-User-Id", required = true) UUID xUserId,
        @Parameter(name = "lessonId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("lessonId") UUID lessonId,
        @Parameter(name = "NewWordDTO", description = "", required = true) @Valid @RequestBody NewWordDTO newWordDTO
    );


    /**
     * DELETE /lessons/{lessonId} : Delete a lesson
     *
     * @param xUserId  (required)
     * @param lessonId  (required)
     * @return Lesson deleted (status code 204)
     *         or Invalid input (status code 400)
     *         or User ID is not recognized (status code 401)
     *         or Unauthorized access to lesson (status code 403)
     */
    @Operation(
        operationId = "deleteLesson",
        summary = "Delete a lesson",
        tags = { "Lesson" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Lesson deleted"),
            @ApiResponse(responseCode = "400", description = "Invalid input"),
            @ApiResponse(responseCode = "401", description = "User ID is not recognized"),
            @ApiResponse(responseCode = "403", description = "Unauthorized access to lesson")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/lessons/{lessonId}"
    )
    
    ResponseEntity<Void> deleteLesson(
        @NotNull @Parameter(name = "X-User-Id", description = "", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "X-User-Id", required = true) UUID xUserId,
        @Parameter(name = "lessonId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("lessonId") UUID lessonId
    );


    /**
     * GET /lessons/{lessonId}/words : Get all words of a lesson
     *
     * @param xUserId  (required)
     * @param lessonId  (required)
     * @return Words of the lesson (status code 200)
     *         or Invalid input (status code 400)
     *         or User ID is not recognized (status code 401)
     *         or Unauthorized access to lesson (status code 403)
     *         or Lesson not found (status code 404)
     */
    @Operation(
        operationId = "getWordsByLessonId",
        summary = "Get all words of a lesson",
        tags = { "Lesson" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Words of the lesson", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = WordDTO.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid input"),
            @ApiResponse(responseCode = "401", description = "User ID is not recognized"),
            @ApiResponse(responseCode = "403", description = "Unauthorized access to lesson"),
            @ApiResponse(responseCode = "404", description = "Lesson not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/lessons/{lessonId}/words",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<WordDTO>> getWordsByLessonId(
        @NotNull @Parameter(name = "X-User-Id", description = "", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "X-User-Id", required = true) UUID xUserId,
        @Parameter(name = "lessonId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("lessonId") UUID lessonId
    );


    /**
     * PUT /lessons : Update a lesson
     *
     * @param xUserId  (required)
     * @param lessonDTO  (required)
     * @return Lesson updated (status code 204)
     *         or Invalid input (status code 400)
     *         or User ID is not recognized (status code 401)
     *         or Unauthorized access to lesson (status code 403)
     *         or Lesson not found (status code 404)
     *         or Lesson name already exists (status code 409)
     */
    @Operation(
        operationId = "updateLesson",
        summary = "Update a lesson",
        tags = { "Lesson" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Lesson updated"),
            @ApiResponse(responseCode = "400", description = "Invalid input"),
            @ApiResponse(responseCode = "401", description = "User ID is not recognized"),
            @ApiResponse(responseCode = "403", description = "Unauthorized access to lesson"),
            @ApiResponse(responseCode = "404", description = "Lesson not found"),
            @ApiResponse(responseCode = "409", description = "Lesson name already exists")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/lessons",
        consumes = { "application/json" }
    )
    
    ResponseEntity<Void> updateLesson(
        @NotNull @Parameter(name = "X-User-Id", description = "", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "X-User-Id", required = true) UUID xUserId,
        @Parameter(name = "LessonDTO", description = "", required = true) @Valid @RequestBody LessonDTO lessonDTO
    );

}