# backend/target/classes/openapi.yaml
openapi: 3.0.0
info:
  title: Word API
  version: 1.0.0
servers:
  - url: /api
...
paths:
...
  /lessons:
    post:
      summary: Create a new lesson
      tags:
        - Lesson
      operationId: createLesson
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewLessonDTO'
      responses:
        '200':
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonDTO'
        '400':
          description: Invalid input
        '401':
          description: User ID is not recognized
        '409':
          description: Lesson name already exists

    put:
      summary: Update a lesson
      tags:
        - Lesson
      operationId: updateLesson
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonDTO'
      responses:
        '204':
          description: Lesson updated
        '400':
          description: Invalid input
        '401':
          description: User ID is not recognized
        '403':
          description: Unauthorized access to lesson
        '404':
          description: Lesson not found
        '409':
          description: Lesson name already exists
  /lessons/{lessonId}:
    delete:
      summary: Delete a lesson
      tags:
        - Lesson
      operationId: deleteLesson
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Lesson deleted
        '400':
          description: Invalid input
        '401':
          description: User ID is not recognized
        '403':
          description: Unauthorized access to lesson
  /lessons/{lessonId}/copy:
    post:
      summary: Copy a shared lesson
      tags:
        - Lesson
      operationId: copySharedLesson
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lesson copied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonDTO'
        '400':
          description: Invalid input
        '401':
          description: User ID is not recognized
        '403':
          description: Unauthorized access to lesson
        '404':
          description: Lesson not found
        '409':
          description: Lesson name already exists
  /lessons/{lessonId}/words:
    get:
      summary: Get all words of a lesson
      tags:
        - Lesson
      operationId: getWordsByLessonId
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Words of the lesson
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WordDTO'
        '400':
          description: Invalid input
        '401':
          description: User ID is not recognized
        '403':
          description: Unauthorized access to lesson
        '404':
          description: Lesson not found

    post:
      summary: Create new word in a lesson
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      tags:
        - Lesson
      operationId: createWord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewWordDTO'
      responses:
        '200':
          description: Word created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WordDTO'
        '400':
          description: Invalid input
        '401':
          description: User ID is not recognized
        '403':
          description: Unauthorized access to lesson
        '404':
          description: Lesson not found
...
components:
  schemas:
    WordDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        question:
          type: string
          minLength: 1
          maxLength: 40
          pattern: '^[\p{L}0-9 ]+$'
          example: "1 black dog"
        answer:
          type: string
          minLength: 1
          maxLength: 40
          pattern: '^[\p{L}0-9 ]+$'
          example: "1 černý pes"
        score:
          type: integer
      required:
        - id
        - question
        - answer
        - score
    NewWordDTO:
      type: object
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 40
          pattern: '^[\p{L}0-9 ]+$'
          example: "1 black dog"
        answer:
          type: string
          minLength: 1
          maxLength: 40
          pattern: '^[\p{L}0-9 ]+$'
          example: "1 černý pes"
      required:
        - question
        - answer
...
    LessonDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lessonName:
          type: string
          minLength: 2
          maxLength: 30
          pattern: '^[\p{L}0-9 ]+$'
          example: "Podstatná jména"
        description:
          type: string
          minLength: 3
          maxLength: 60
          pattern: '^[\p{L}0-9 ]+$'
          example: "V lekci jsou běžná podstatná jména"
        language:
          type: string
          minLength: 2
          maxLength: 30
          pattern: '^[\p{L}0-9 ]+$'
          example: "AJ"
        score:
          type: integer
          format: flow
          example: 3,5
        wordCount:
          type: integer
          format: int32
          example: 15
        shared:
          type: boolean
      required:
        - id
        - lessonName
        - description
        - language
        - score
        - wordCount
        - shared
    NewLessonDTO:
      type: object
      properties:
        lessonName:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[\p{L}0-9 ]+$'
          example: "Podstatná jména"
        description:
          type: string
          minLength: 2
          maxLength: 60
          pattern: '^[\p{L}0-9 ]+$'
          example: "V lekci jsou běžná podstatná jména"
        language:
          type: string
          minLength: 2
          maxLength: 30
          pattern: '^[\p{L}0-9 ]+$'
          example: "AJ"
        shared:
          type: boolean
      required:
        - lessonName
        - description
        - language
        - shared
...
  parameters:
    UserIdHeader:
      name: X-User-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid